+++ app.js
@@ -1,5 +1,5 @@
-import { autoRolling } from "./rolling_banner.js";
-import { initPress } from "./random_news.js";
+import { autoRolling } from "./src/views/rolling_views.js";
+import { initPress } from "./src/views/grid_views.js";
 
 function initApp() {
     autoRolling();
