+++ index.html
@@ -113,7 +113,7 @@
             </div>
         </div>
     </body>
-    <script type="module" src="./utils.js"></script>
-    <script type="module" src="./events.js"></script>
+    <script type="module" src="./src/utils.js"></script>
+    <script type="module" src="./src/events.js"></script>
     <script type="module" src="./app.js"></script>
 </html>
